feat: websocket implementation & events routings added (#8)
* fix: fixed deploy

* fix: database

* feat: added events route and web socket

+++ backend/adapters/repositories-integration/src/main/kotlin/com/y9vad9/pomodoro/backend/repositories/integration/tables/TimersEventsTable.kt
@@ -6,7 +6,7 @@ import org.jetbrains.exposed.sql.Table
 import org.jetbrains.exposed.sql.select
 
 internal object TimersEventsTable : Table() {
-    val EVENT_ID = integer("timer_event_id")
+    val EVENT_ID = long("timer_event_id")
         .autoIncrement()
 
     val TIMER_ID = integer("timer_id")
